Start of NN - Waveforms SDK integration
The "NN on Memristors" folder is designed to be the start of combining our code for controlling the Waveforms SDK (initially Python Waveforms test folder) and with pytorch.

Current experiment is to get the power supply working in the same file as the DIO pins. It is working in ANALOGIO_AnalogDiscovery2_Power.py, but in the more rudimentary version. It would be more usable if brought to a higher level, like in AD2-mux-test. You can see that this file controls the DIO pins but not the power supplies in its current state.